Persistent checkpoints.
Does not allow for syscall buffering (and currently does no checking for
if the record was with or without; thus will fail badly if not used in
tandem with `record -n`)

Using the persistent checkpoints right now:

`write-checkpoints` - writes all the checkpoints the user has added to
disk. This also deletes any checkpoints on disk, that is currently not
on the list, when using `info checkpoints`. To delete all checkpoints,
do a `rr replay` and enter `write-checkpoints`. Not great, but that's
how it works right now.

`load-serialized-checkpoints`: loads and sets the checkpoints, makes
them show up in `info checkpoints`. From there, they should behave like
normal GDB Checkpoints, i.e., `restart <checkpoint id>`.

If you have persistent checkpoints at event A, B, C, when
replaying `rr replay -g (>=A)` rr will start from A, and the same goes
for B and C. Due to having to test this feature out, another flag has
been added temporarily, which is the `-b` flag, which behaves like `-g`
but it starts from the beginnning, regardless if persistent checkpoints
exist or not.

Many optimizations is (probably) missing. Currently serializes
redundant/duplicated mappings, taking up unnecessary space.

Known Bugs:
When reverse-continuing over a persistent checkpoint boundary, strange
results are seen. It doesn't crash or burn, but, it seems to "lag"
across it, seemingly replaying that event several times, until finally
pulling through.

The serialization of `FileMonitor` is done on a whim, and it might be
right or it might be wrong. Probably needs input here.

Some static `TraceStream.cc` functions has been moved to
`stream_util.hh/cc`.