Persistent checkpointing
Added persistent checkpoint schema for capnproto rr_pcp.capnp,
as well a compile command for it in CMakeLists.txt, that works
just like the normal trace schema

CheckpointInfo and MarkData types works as intermediaries between a
serialized checkpoint and a deserialized one. MarkData is used for
copying the contents of Mark, InternalMark, ProtoMark and it's various
data into, for serialization as well when deserializing, to reconstruct
those types.

The reasoning for this is to not intrude in their interface and possibly
break some guarantees or invariants they provide. If something goes wrong now, it's
constrained only to persistent checkpointing not reconstructing those
data types correctly.

GDB spawned by RR now has 2 additional commands, write-checkpoints and
load-checkpoints for which load_checkpoint and serialize_checkpoint performs the
work in ReplaySession.

rr create-checkpoints command added which create persistent checkpoints
on a specified interval.