Added PCP functionality to Replay & ReRun commands
Replay sessions now get spawned from persistent checkpoints if they
exist on disk.

-g <evt>
will look for a event prior to <evt> (closest) that is a persistent
checkpoint and spawn from that.

-f <pid>
When following a <pid>, the replay command will look for it's origin
event, and if that is after a pcp, the command will pick the closest pcp
and spawn from that and continue from there.

Added the --ignore-pcp flag for when one do not want to use pcps to
spawn from. Using pcps is default on.