Add move constructors to BreakStatus
This bug bites me every time I'm testing new things and I add a new field to this and have forgotten that it defines copy constructors but not defaulted move constructors.

Also, this unnecessarily copies itself every time this gets returned out of a function, either directly, or via one of the composite types that contains it, like ReplayResult.