Make RR lib'able
v0.1 of RR as a library. Makes it possible to statically link against librr.a. RR_AS_LIBRARY as a compile definition, can be used to turn off things, that we don't want in the library (potentially logging to stdout, for instance).

Coming commits will turn off the emergency debugger with this